backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
langgraph
tavily-python
beautifulsoup4
requests
python-dotenv
pydantic
sse-starlette
tiktoken
httpx
scikit-learn
numpy
